// src/app/purchase/purchase-manage/price.type.ts
export type PriceObject = {
  cessAmount?:             number;
  cgstAmount?:             number;
  discountAmount?:         number;
  grossAmount?:            number;
  igstAmount?:             number;
  netTotalAmount?:         number;
  otherExpense?:           number;
  roundedNetTotalAmount?:  number;
  roundOff?:               number;
  sgstAmount?:             number;
  subTotal?:               number;
  addnlCessAmount?:        number;
  transExpense?:           number;
  remark?:                 string;
  };
  